refactor: 각 패키지 종류에 맞게 이동
