+++ dev-route/src/main/java/com/teamdevroute/devroute/user/controller/UserController.java
@@ -1,6 +1,8 @@
-package com.teamdevroute.devroute.user;
+package com.teamdevroute.devroute.user.controller;
 
 import com.teamdevroute.devroute.global.auth.LoginUserInfo;
+import com.teamdevroute.devroute.user.dto.UserLoginRequest;
+import com.teamdevroute.devroute.user.service.UserService;
 import com.teamdevroute.devroute.user.domain.CustomUserDetails;
 import com.teamdevroute.devroute.user.dto.UserCreateRequest;
 import com.teamdevroute.devroute.user.dto.UserCreateResponse;
