+++ dev-route/src/test/java/com/teamdevroute/devroute/user/UserServiceTest.java
@@ -2,19 +2,12 @@ package com.teamdevroute.devroute.user;
 
 import com.teamdevroute.devroute.user.domain.User;
 import com.teamdevroute.devroute.user.dto.UserCreateRequest;
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
+import com.teamdevroute.devroute.user.service.UserService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
-
-import java.util.HashMap;
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
